refactor(roadmap-generation): update roadmap generation route to /roadmaps/generate
