+++ apps/web/app/(full-layout)/(home)/_components/hero-section.tsx
@@ -81,7 +81,7 @@ export function HeroSection() {
             <AnimatedIconSwap icon={ArrowRight} hoverIcon={ArrowRight02FreeIcons} />
           </Button>
           <Button variant="outline" size="lg" className="group/btn rounded-full">
-            <Link href={'/generate-roadmap' as Route<string>}>Create your roadmap</Link>
+            <Link href={'/roadmaps/generate' as Route<string>}>Create your roadmap</Link>
             <AnimatedIconSwap icon={ArrowRight} hoverIcon={ArrowRight02FreeIcons} />
           </Button>
         </div>
