+++ apps/web/app/(full-layout)/roadmaps/generate/_components/generation-loading.tsx
