+++ apps/web/app/(full-layout)/roadmaps/generate/_types/onboarding.ts
