+++ apps/web/app/(full-layout)/roadmaps/generate/_validations/ai-roadmap.schema.ts
