+++ apps/web/services/roadmap-service.ts
@@ -1,7 +1,8 @@
-import {
-  type GenerateRoadmapPayload,
-  type GenerateRoadmapResponse,
-} from '@/app/(full-layout)/generate-roadmap/_types/onboarding';
+import type {
+  GenerateRoadmapPayload,
+  GenerateRoadmapResponse,
+} from '@/app/(full-layout)/roadmaps/generate/_types/onboarding';
+
 import { ENDPOINTS } from '@/constants/endpoints';
 import { axiosInstance } from '@/lib/axios-instance';
